Crash Dump
Is anyone else having this problem or is this an isolated issue (on Steam) with certain Gamers like myself? I noticed multiple threads have been created expressing Gamers having problems - but I'm not sure if it's related to my ability to launch the actual game; followed by "continuing the game."

Is this related to our previous mods? 
Do the mods make the game not compatible with the new update?
Does your error say to send the developer of the game your executable file?